redistributed methods from CameraHandler to CameraChest and CameraObject. redistrubited methods from CameraHandlerTest to CameraChestTest and CameraObjectTest
